ISSUE #3650 - Added fetch the ticket details when navigating to a ticket's detail
