+++ frontend/src/v5/ui/routes/viewer/tickets/ticketsList/ticketsList.component.tsx
@@ -58,7 +58,7 @@ export const TicketsList = ({ tickets }: TicketsListProps) => {
 
 	const onTicketClick = (ticket: ITicket) => {
 		if (ticketIsSelected(ticket)) {
-			contextValue.setCardView(TicketsCardViews.Details, { ticket });
+			contextValue.setCardView(TicketsCardViews.Details, { ticketId: ticket._id });
 		} else {
 			setSelectedTicket(ticket);
 		}
